feat: when remote installing, check if we can just update from git first

+++ remote-git-install.sh
@@ -1,28 +1,48 @@
 #!/usr/bin/env bash
 
-# install pacapt to install git
-wget -O /tmp/pacapt https://github.com/icy/pacapt/raw/ng/pacapt
-chmod 755 /tmp/pacapt
+SKIPINSTALL=0
 
-# install git
-sudo /tmp/pacapt -S --noconfirm git
+# check to see if it's already cloned
+if [ -d "$HOME/.dotfiles" ]; then
+    if [ -d "$HOME/.dotfiles/.git" ]; then
+        SKIPINSTALL=1
 
-# remove pacapt
-rm -rf /tmp/pacapt
+        printf "Repository already exists, pulling latest changes\n\n"
 
-# remove existing directory
-rm -rf "$HOME/.dotfiles"
+        cd "$HOME/.dotfiles"
 
-# clone the repository from https
-git clone --branch master https://github.com/RyanTheAllmighty/dotfiles.git "$HOME/.dotfiles"
+        # stash any changes, pull latest changes, then pop the stash back
+        git stash
+        git pull
+        git stash pop
+    fi
+fi
 
-cd "$HOME/.dotfiles"
+if [ $SKIPINSTALL = 0 ]; then
+    # install pacapt to install git
+    wget -O /tmp/pacapt https://github.com/icy/pacapt/raw/ng/pacapt
+    chmod 755 /tmp/pacapt
 
-if [[ `git remote get-url origin` = "https://github.com/RyanTheAllmighty/dotfiles.git" ]]; then
-    printf "\n\nUsing https git url. Please consider using the ssh one\n\n"
-    printf "git remote set-url origin git@github.com:RyanTheAllmighty/dotfiles.git\n\n"
-    printf "Press enter to continue\n"
-    read
+    # install git
+    sudo /tmp/pacapt -S --noconfirm git
+
+    # remove pacapt
+    rm -rf /tmp/pacapt
+
+    # remove existing directory
+    rm -rf "$HOME/.dotfiles"
+
+    # clone the repository from https
+    git clone --branch master https://github.com/RyanTheAllmighty/dotfiles.git "$HOME/.dotfiles"
+
+    cd "$HOME/.dotfiles"
+
+    if [[ `git remote get-url origin` = "https://github.com/RyanTheAllmighty/dotfiles.git" ]]; then
+        printf "\n\nUsing https git url. Please consider using the ssh one\n\n"
+        printf "git remote set-url origin git@github.com:RyanTheAllmighty/dotfiles.git\n\n"
+        printf "Press enter to continue\n"
+        read
+    fi
 fi
 
 # run the installer
